requirements_dev.txt
flake8==3.5.0
tox==3.5.2
pip==18.1
bumpversion==0.5.3
wheel==0.32.1
watchdog==0.9.0
coverage==4.5.1
Sphinx==1.8.1
twine==1.12.1
pytest==3.8.2
pytest-runner==4.2
django==2.1
factory-boy==2.11.1
django-choices==1.6.2
-e git+https://github.com/maykinmedia/chemistry-cmislib.git@0.7.dev5#egg=cmislib
django-webtest==1.9.4
unittest-xml-reporting==2.5.1
invoke==1.2.0
